refactor: separate Daily from SingleCardLayout
